src/backend/agent/requirements.txt
langchain==0.3.26
langchain-community==0.3.26
langchain-openai==0.3.27
faiss-cpu==1.11.0
openai==1.93.0
tiktoken==0.9.0
pymongo==4.13.2
langchain-ollama==0.3.3
ragas==0.2.15
fastapi==0.116.0
uvicorn==0.35.0
PyYAML==6.0.2
sentence-transformers==5.0.0
ddgs==9.2.3
pydantic==2.11.7
langgraph==0.5.3
mcp[cli]==1.13.1
